requirements.txt
polars==1.27.1
psycopg2-binary==2.9.10
pyarrow==19.0.1
pandas==2.2.3
